Suppress facebook-unused-include-check for boost/thread/thread.hpp in AtomicMapTest.cpp
Summary:
Added `// NOLINT` comment to suppress the facebook-unused-include-check warning for the boost/thread/thread.hpp include in AtomicMapTest.cpp.

The linter flagged this as an "umbrella header" that should be replaced with more specific includes. However, the boost::thread class is essential for the concurrent testing in this file (creating 1000 threads for testing AtomicMap operations), and the include is actually needed.

This fix follows the established pattern in the redex codebase:
- fbandroid/native/redex/libredex/WorkQueue.h (line 6)
- fbandroid/native/redex/test/unit/ConcurrentHashtableTest.cpp (line 18)

Both files use the same `// NOLINT` comment to suppress the identical warning for the same boost/thread/thread.hpp include.

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243011139-47a5c4b7-82a5-45f6-807b-bf969d6b0676)

Reviewed By: xuhdev

Differential Revision: D85797555

fbshipit-source-id: 50fedc9346b65ff0a6511f73a38e0e8ab06a539b