Fix facebook-unused-include-check warning in DexGrep.cpp
Summary:
Fixed the `facebook-unused-include-check` linter warning for `<getopt.h>` in DexGrep.cpp by adding the `donotremove` pragma.

The linter was flagging `<getopt.h>` as an "umbrella header" that should be replaced with more specific headers. However, `<getopt.h>` is the correct POSIX header for the symbols used in this file:
- `getopt_long()` function
- `struct option`
- `no_argument` constant
- `optind` variable

According to the Include Cleaner documentation (https://fburl.com/include), the `donotremove` pragma is the appropriate way to exclude includes from analysis when they are necessary but flagged by the linter.

This change resolves T243011254.

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243011254-baba0dd2-eb90-417a-8d22-609f13c3c15b)

Reviewed By: xuhdev

Differential Revision: D85800546

fbshipit-source-id: 41a4d3f27ca911066b018b85ba7b24fe747de9a1