Fix B903 lint warning in dump_classes_from_hprof.py by adding __slots__
Summary:
Added __slots__ declaration to the SimpleSegment class to address the flake8-bugbear B903 warning.

The B903 warning indicates that data classes should either be immutable or use __slots__ to save memory. Since SimpleSegment is a simple data class with only two attributes (heap_tag and object_id) and is subclassed by HprofRoot and HprofObject, adding __slots__ is the appropriate solution.

This change:
- Reduces memory overhead by preventing the creation of __dict__ for each instance
- Maintains backward compatibility as the class interface remains unchanged
- Follows the pattern used in similar fixes across the codebase

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243801751-e4eeca5f-cdb2-456b-bc96-7dc37289effe)

Reviewed By: litiangong

Differential Revision: D86155395

fbshipit-source-id: 09148a05ede4c07376cf52fade6ae28860f6cc45