Fix facebook-unused-include-check in VirtScopeTest.cpp
Summary:
Fixed facebook-unused-include-check clang-tidy violation by removing the unnecessary #include "Creators.h" from VirtScopeTest.cpp.

The issue was that Creators.h is an umbrella header that was included in the test file but not actually used. The test file uses create_scope_* functions which are declared in VirtScopeHelper.h (already included) and implemented in VirtScopeHelper.cpp (which includes Creators.h where it's actually needed).

All necessary types and functions remain available through the other includes:
- DexClass.h provides DexClass, DexMethod, DexType, etc.
- VirtScopeHelper.h provides create_scope_* function declarations
- Other includes provide the remaining necessary types

This change follows the clang-tidy rule that flags includes that "only serve as an umbrella to other headers" and reduces compilation dependencies.

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243011221-c6d9e772-096b-49a0-94a1-dc1b32caa7aa)

Reviewed By: litiangong

Differential Revision: D86164796

fbshipit-source-id: c9ce8fd3836b740a89db6f8d39202d0edda28778