Add option to skip R_CONST instr insertion in MaterializeResourcesConstants
Summary:
High level goal of this stack is to enable OptimizeResourcesPass on WhatsApp which would allow resource ID compaction (decreases size of arsc file). [Document explaining this](https://docs.google.com/document/d/1As2XCLfdEujG7hoGMhBpsyC0wrGJPSNaty7UFb0oLqU/edit?usp=sharing)

We introduced MaterializeResourceConstantsPass so that we can remap the StringPack resource IDs, as mentioned in D86112729. However, MaterializeResourceConstantsPass does some extra other things that we do not want to enable quite yet, so we introduce a flag to hide that functionality.

Note that MRCP is not actually used at this point, so there is no material change to Redex behaviour.

Reviewed By: wsanville

Differential Revision: D86113832

fbshipit-source-id: bfeedb3e8674dab0ffbfdc53e7a3d72891d569b0