Check existence of left out aputs before patching array literals
Summary:
# Context

The analysis in ReducedArrayLiteralsPass tracks the flow of an array literal construction starting with a `new-array` followed by a number of `aput`s. If the array value "escapes", the analysis registers it as an escaped array literal. However, when the new array being constructed flows through control flow branches, one of the branches might not have an instruction that registers the "escape" of the array value. As a result, the analysis will ignore the `aput`s in the non-escaping branch but still register the ones on the escaping branch.

The new `ReduceArrayLiteralsTest.conditional_escape2` illustrates this scenario. When the analysis left out the `aput` in the `:skip_aput` branch, the pass still replaces the array construction instructions with `filled-new-array` on the other branch leaving the `aput` with no incoming array value.

# Solution

In the `ReduceArrayLiterals.cpp` file, we add a new helper `has_left_out_aputs` that checks if there are any `aput` instructions consuming the `new-array` instruction that are not captured by the analysis or have been "left out". This function will be used to determine whether the array literal can be safely optimized.

The updated code is designed to improve the reliability and effectiveness of the ReduceArrayLiterals pass in optimizing array literals.

Reviewed By: agampe

Differential Revision: D88122122

fbshipit-source-id: 52859f7831045d44ffa19f13c0a8ac65ace6a071