Fix bug in chain and dom violations
Summary:
In the case where a source block with an undefined value is preceded by a source block with a hot value (and it's within the same block and not separated by a throw), we would previously not label this as an error.

However, this is incorrect because in the code we state that we treat undefined values as 0's. Thus, it should be an error.

This fixes the issue

Reviewed By: agampe

Differential Revision: D89393598

fbshipit-source-id: 23d2832d3d29a1f4cab6c59c855fc121ebd8f8eb