SwitchEquivFinder test case for mix of if and switch statements
Summary: SwitchEquivFinder was meant to be a tool for gracefully handling the many ways in which a conceptual switch could get implemented - be it an actual switch, if statements, or a combination of the two. This was not shown in any test cases, so this diff simply enshrines an example of that and verifies the basic understanding.

Reviewed By: xuhdev

Differential Revision: D89745393

fbshipit-source-id: e0f2c6525a43260eee950a203007c89581fbc589